Fix loading of documents with relative paths
HTML anchors with an href that does not start with a `/`, i.e. that have
a relative path, need to be prefixed with `/` to properly join them to
the base host.

An example can be seen on https://unobtanium.de. The links are relative
to the root directory. Without this patch, the browser crashes when trying
to load them.